+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [3.1.0] - 2025-06-22
 ### Added
 - Readonly attribute
 
@@ -64,7 +66,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add initial files
 - Initial commit
 
-[Unreleased]: https://github.com/HyagoOliveira/attributes/compare/3.0.0...main
+[Unreleased]: https://github.com/HyagoOliveira/attributes/compare/3.1.0...main
+[3.1.0]: https://github.com/HyagoOliveira/attributes/tree/3.1.0/
 [3.0.0]: https://github.com/HyagoOliveira/attributes/tree/3.0.0/
 [2.1.0]: https://github.com/HyagoOliveira/attributes/tree/2.1.0/
 [2.0.1]: https://github.com/HyagoOliveira/attributes/tree/2.0.1/